refactor(agent/tools): review 6 项修复
1. AgentContext 加 persona_id，recall.py 从 context 取代替 header_vars
   删除 _get_persona_id() 硬编码 "chiwei" 的 fallback
2. search_images 拆分到 image_search.py（search.py 511→380 行）
3. Prometheus _names_to_collectors 加注释说明无官方替代
4. _fetch_content 不再 mutate 输入 dict
5. search_web 去掉与 _rerank_chunks 内部 fallback 冗余的外层 catch
6. history.py 内联查询下沉到 data/queries.py
   新增 find_context_messages_for_anchors + find_group_members
